user deleteのアソシエーション対応

+++ controllers/usersController.go
@@ -305,6 +305,11 @@ func UsersDelete(c *fiber.Ctx) error {
 			log.Printf("db error: %v", errLicense)
 			return fmt.Errorf("db error: %v", errLicense)
 		}
+		errSchool := tx.Table("schools").Where("user_id = ?", user.Id).Delete("").Error
+		if errSchool != nil {
+			log.Printf("db error: %v", errSchool)
+			return fmt.Errorf("db error: %v", errSchool)
+		}
 
 		// user情報の削除
 		err = tx.Delete(user).Error
